Refactor BladeSchemaForm
