+++ src/components/SchemaForm/BladeColorDropDown.jsx
@@ -13,7 +13,7 @@ const useStyles = makeStyles((theme) => ({
   FormRow: {
     display: 'flex',
     flexDirection: 'row',
-    border: '1px solid',
+    border: '1px solid rgba(255, 255, 255, 0.23)',
     borderRadius: '10px',
     '@media (max-width: 580px)': {
       flexDirection: 'column',
