+++ src/components/SchemaForm/BladeEffectSchemaForm.jsx
@@ -15,6 +15,8 @@ import {
   FormControl,
 } from '@material-ui/core';
 import SettingsIcon from '@material-ui/icons/Settings';
+import useStore from '../../utils/apiStore';
+import BladeColorDropDown from './BladeColorDropDown';
 import BladeBoolean from './BladeBoolean';
 import BladeSelect from './BladeSelect';
 import BladeSlider from './BladeSlider';
@@ -27,18 +29,33 @@ const useStyles = makeStyles({
   },
 });
 
-const BladeSchemaFormNew = (props) => {
+const BladeEffectSchemaForm = (props) => {
   const {
+    effects,
+    virtual,
     schema,
     model,
+    virtual_id,
+    selectedType,
     colorMode = 'picker',
+    colorKeys = [],
     boolMode = 'switch',
     boolVariant = 'outlined',
     selectVariant = 'outlined',
     sliderVariant = 'outlined',
-    onModelChange = (e) => e,
   } = props;
-
+  const pickerKeys = [
+    'color',
+    'background_color',
+    'color_lows',
+    'color_mids',
+    'color_high',
+    'strobe_color',
+    'lows_colour',
+    'mids_colour',
+    'high_colour',
+    ...colorKeys,
+  ];
 
   const classes = useStyles();
 
@@ -48,6 +65,8 @@ const BladeSchemaFormNew = (props) => {
   const [_selectVariant, _setSelectVariant] = useState(selectVariant);
   const [_sliderVariant, _setSliderVariant] = useState(sliderVariant);
   const [_colorMode, _setColorMode] = useState(colorMode);
+  const updateVirtualEffect = useStore((state) => state.updateVirtualEffect);
+  const getVirtuals = useStore((state) => state.getVirtuals);
 
   const handleClickOpen = () => {
     setOpen(true);
@@ -56,6 +75,14 @@ const BladeSchemaFormNew = (props) => {
   const handleClose = () => {
     setOpen(false);
   };
+
+  const handleEffectConfig = (virtual_id, config) => updateVirtualEffect(virtual_id, {
+    virtId: virtual_id,
+    type: selectedType,
+    config,
+  }).then(() => {
+    getVirtuals();
+  });
 
   return (
     <div className={classes.bladeSchemaForm}>
@@ -70,14 +97,22 @@ const BladeSchemaFormNew = (props) => {
           <SettingsIcon />
         </Fab>
       )}
+      {pickerKeys && pickerKeys.map(
+        (k, i) => model && Object.keys(model).indexOf(k) !== -1 && (
+          <BladeColorDropDown
+            virtual={virtual}
+            effects={effects}
+            selectedType={selectedType}
+            model={model}
+            key={k}
+            index={i}
+            type={_colorMode === 'select' ? 'text' : 'color'}
+            clr={k}
+          />
+        ),
+      )}
 
-
-      {schema && schema.properties && Object.keys(schema.properties).map((s, i) => {
-        let permitted = true
-        if (schema.permitted_keys && schema.permitted_keys.indexOf(s) === -1) {
-          permitted = false
-        }
-        
+      {Object.keys(schema.properties).map((s, i) => {
         switch (schema.properties[s].type) {
           case 'boolean':
             return (
@@ -85,107 +120,85 @@ const BladeSchemaFormNew = (props) => {
                 type={_boolMode}
                 variant={_boolVariant}
                 key={i}
+                index={i}
                 model={model}
                 model_id={s}
-                required={schema.required && schema.required.indexOf(s) !== -1}
-                style={{ margin: '0.5rem 0', flexBasis: '48%' }}
                 schema={schema.properties[s]}
+                hideDesc={true}
                 onClick={(model_id, value) => {
                   const c = {};
                   c[model_id] = value;
-                  return onModelChange(c);
+                  return handleEffectConfig(virtual_id, c);
                 }}
               />
             );
           case 'string':
-            return <BladeSelect
-              model={model}
-              disabled={!permitted}
-              style={{ margin: '0.5rem 0', width: '48%' }}
-              variant={_selectVariant}
-              schema={schema.properties[s]}
-              required={schema.required && schema.required.indexOf(s) !== -1}
-              model_id={s}
-              key={i}
-              onChange={(model_id, value) => {
-                const c = {};
-                c[model_id] = value;
-                return onModelChange(c);
-              }}
-            />
+            return schema.properties[s].enum && pickerKeys.indexOf(s) === -1 ? (
+              <BladeSelect
+                model={model}
+                variant={_selectVariant}
+                schema={schema.properties[s]}
+                wrapperStyle={{width: '48%'}}
+                model_id={s}
+                key={i}
+                index={i}
+                onChange={(model_id, value) => {
+                  const c = {};
+                  c[model_id] = value;
+                  return handleEffectConfig(virtual_id, c);
+                }}
+              />
+            ) : (
+              pickerKeys.indexOf(s) === -1 && (
+                <BladeColorDropDown
+                  selectedType={selectedType}                  
+                  index={i}
+                  model={model}
+                  type="colorNew"
+                  clr="blade_color"
+                  key={i}
+                />
+              )
+            );
 
           case 'number':
             return (
               <BladeSlider
                 variant={_sliderVariant}
-                disabled={!permitted}
                 key={i}
+                index={i}
+                hideDesc={true}
                 model_id={s}
                 model={model}
-                required={schema.required && schema.required.indexOf(s) !== -1}
                 schema={schema.properties[s]}
                 onChange={(model_id, value) => {
                   const c = {};
                   c[model_id] = value;
-                  return onModelChange(c);
+                  return handleEffectConfig(virtual_id, c);
                 }}
               />
             );
 
           case 'integer':
-            return <BladeSlider
-              variant={_sliderVariant}
-              disabled={!permitted}
-              step={1}
-              key={i}
-              model_id={s}
-              model={model}
-              required={schema.required && schema.required.indexOf(s) !== -1}
-              schema={schema.properties[s]}
-              textfield={true}
-              style={{ margin: '0.5rem 0', width: '48%' }}
-              onChange={(model_id, value) => {
-                const c = {};
-                c[model_id] = value;
-                return onModelChange(c);
-              }}
-            />
-          case 'int':
-            return schema.properties[s]?.enum?.length > 10 ? <BladeSlider
-              variant={_sliderVariant}
-              disabled={!permitted}
-              marks={schema.properties[s]?.enum}
-              step={null}
-              key={i}
-              model_id={s}
-              model={model}
-              required={schema.required && schema.required.indexOf(s) !== -1}
-              schema={schema.properties[s]}
-              textfield={false}
-              style={{ margin: '0.5rem 0', width: '48%' }}
-              onChange={(model_id, value) => {
-                const c = {};
-                c[model_id] = value;
-                return onModelChange(c);
-              }}
-            /> : <BladeSlider
-              variant={_sliderVariant}
-              disabled={!permitted}
-              marks={schema.properties[s]?.enum}
-              step={null}
-              key={i}
-              model_id={s}
-              model={model}
-              required={schema.required && schema.required.indexOf(s) !== -1}
-              schema={schema.properties[s]}
-              textfield={false}
-              style={{ margin: '0.5rem 0', width: '48%' }}
-              onChange={(model_id, value) => {
-                const c = {};
-                c[model_id] = value;
-                return onModelChange(c);
-              }}
-            />
+            return (
+              <BladeSlider
+                variant={_sliderVariant}
+                step={1}
+                key={i}
+                index={i}
+                hideDesc={true}
+                model_id={s}
+                model={model}
+                schema={schema.properties[s]}
+                style={{ margin: '0.5rem 0', flexBasis: '47%'}}
+                onChange={(model_id, value) => {
+                  const c = {};
+                  c[model_id] = value;
+                  return handleEffectConfig(virtual_id, c);
+                }}
+              />
+            );
+
           default:
             return (
               <>
@@ -281,14 +294,17 @@ const BladeSchemaFormNew = (props) => {
   );
 };
 
-BladeSchemaFormNew.propTypes = {
+BladeEffectSchemaForm.propTypes = {
   colorMode: PropTypes.oneOf(['picker', 'select']),
   boolMode: PropTypes.oneOf(['switch', 'checkbox', 'button']),
   boolVariant: PropTypes.oneOf(['outlined', 'contained', 'text']),
   selectVariant: PropTypes.string, // outlined | any
   sliderVariant: PropTypes.string, // outlined | any
+  colorKeys: PropTypes.array,
   schema: PropTypes.object.isRequired,
   model: PropTypes.object.isRequired,
+  virtual_id: PropTypes.string.isRequired,
+  selectedType: PropTypes.string.isRequired,
 };
 
-export default BladeSchemaFormNew;
+export default BladeEffectSchemaForm;
